# Efficient parallel CPU utilization for solving