Replaces the single channel based explore iterator with generator functions that can be split for entirely isolated concurrent execution without overlap.

Prior to this, when running on multiple CPUs, the majority of the goroutines were starved on contention waiting for new data from the explore iterator:
![profile002](https://user-images.githubusercontent.com/40650/148651019-a02d8d14-97fd-4c24-a52a-cb3ccb20aa51.png)

After the change, the majority of the CPU time is back to hashing:
![profile003](https://user-images.githubusercontent.com/40650/148651055-9c6fbb49-fc7f-4cdc-815c-3bb9042eb724.png)

On my laptop (2018 MacBook Pro, intel 4 cores hyperthreaded), this results in a ~45% performance increase for the core solve .
```
name             old time/op    new time/op    delta
SolveParallel       658ns ± 1%     585ns ± 1%  -11.05%  (p=0.008 n=5+5)
SolveParallel-2     388ns ± 3%     301ns ± 2%  -22.40%  (p=0.008 n=5+5)
SolveParallel-4     250ns ± 8%     190ns ±10%  -24.02%  (p=0.008 n=5+5)
SolveParallel-8     283ns ± 2%     193ns ± 2%  -31.93%  (p=0.008 n=5+5)

name             old speed      new speed      delta
SolveParallel     527MB/s ± 1%   593MB/s ± 1%  +12.42%  (p=0.008 n=5+5)
SolveParallel-2   894MB/s ± 3%  1152MB/s ± 1%  +28.82%  (p=0.008 n=5+5)
SolveParallel-4  1.39GB/s ± 8%  1.83GB/s ±10%  +31.66%  (p=0.008 n=5+5)
SolveParallel-8  1.23GB/s ± 2%  1.80GB/s ± 2%  +46.93%  (p=0.008 n=5+5)
```

UPDATE: Testing the same on my partner's 2020 M1 MacBook Air reveals even more significant improvements.
```
name             old time/op    new time/op    delta
SolveParallel       266ns ± 0%     239ns ± 0%   -10.26%  (p=0.008 n=5+5)
SolveParallel-2     179ns ± 1%     123ns ± 0%   -31.06%  (p=0.008 n=5+5)
SolveParallel-4     140ns ± 3%      64ns ± 0%   -54.36%  (p=0.016 n=5+4)
SolveParallel-8     145ns ± 3%      45ns ± 0%   -69.02%  (p=0.008 n=5+5)

name             old speed      new speed      delta
SolveParallel    1.31GB/s ± 0%  1.45GB/s ± 0%   +11.44%  (p=0.008 n=5+5)
SolveParallel-2  1.94GB/s ± 1%  2.81GB/s ± 0%   +45.02%  (p=0.008 n=5+5)
SolveParallel-4  2.48GB/s ± 3%  5.44GB/s ± 0%  +119.04%  (p=0.008 n=5+5)
SolveParallel-8  2.40GB/s ± 3%  7.75GB/s ± 0%  +222.62%  (p=0.008 n=5+5)
```
(Also sheesh, I really need a new laptop.)