@mroth 2023-08-09
Now that I have my own modern laptop, update benchmark comparison:
```
benchstat -filter=".name:SolveParallel" before.txt after.txt
goos: darwin
goarch: arm64
pkg: github.com/mroth/gitbrute
                 │  before.txt  │              after.txt              │
                 │    sec/op    │   sec/op     vs base                │
SolveParallel-12   326.65n ± 0%   20.94n ± 2%  -93.59% (p=0.000 n=10)

                 │  before.txt   │                after.txt                 │
                 │      B/s      │      B/s        vs base                  │
SolveParallel-12   1013.2Mi ± 0%   15801.1Mi ± 2%  +1459.50% (p=0.000 n=10)
```